+++ lib/karafka/pro/routing/features/multiplexing.rb
@@ -1,0 +1,37 @@
+# frozen_string_literal: true
+
+# This Karafka component is a Pro component under a commercial license.
+# This Karafka component is NOT licensed under LGPL.
+#
+# All of the commercial components are present in the lib/karafka/pro directory of this
+# repository and their usage requires commercial license agreement.
+#
+# Karafka has also commercial-friendly license, commercial support and commercial components.
+#
+# By sending a pull request to the pro components, you are agreeing to transfer the copyright of
+# your code to Maciej Mensfeld.
+
+module Karafka
+  module Pro
+    # Namespace for Pro routing enhancements
+    module Routing
+      # Namespace for additional Pro features
+      module Features
+        # Multiplexing allows for creating multiple subscription groups for the same topic inside
+        # of the same subscription group allowing for better parallelism with limited number
+        # of processes
+        class Multiplexing < Base
+          class << self
+            # @param _config [Karafka::Core::Configurable::Node] app config node
+            def pre_setup(_config)
+              # Make sure we use proper unique validator for topics definitions
+              ::Karafka::Contracts::ConsumerGroup.singleton_class.prepend(
+                Patches::Contracts::ConsumerGroup
+              )
+            end
+          end
+        end
+      end
+    end
+  end
+end